Refactoring based on comments
