+++ src/domain/banner/bannerConstants.ts
@@ -1,1 +1,8 @@
+import { LandingPageTextColor } from '../../types';
+
 export const BANNER_SOME_IMAGE = '/images/activities_SoMe-share.jpg';
+
+export const contentBackgroundColorMap: Record<LandingPageTextColor, string> = {
+  BLACK: 'rgba(255, 255, 255, 0.7)',
+  WHITE: 'rgba(0, 0, 0, 0.7)',
+};
